reestructured ticketType choice
